Add state for home fragment, ViewModel and refactor
